change tests to be ran by jest, fulfilling requirement of using Node.js (>=6.5.7)
