+++ .babelrc
@@ -1,0 +1,9 @@
+{
+  "presets": [
+    ["env", {
+      "targets": {
+        "node": "6.5.7"
+      }
+    }]
+  ]
+}
