+++ package.json
@@ -40,6 +40,7 @@
     "underscore": "^1.8.3"
   },
   "devDependencies": {
+    "babel-preset-env": "^1.6.1",
     "chai": "3.3.0",
     "debug": "3.1.0",
     "eslint": "^3.16.0",
@@ -48,7 +49,7 @@
     "eslint-plugin-jsx-a11y": "^4.0.0",
     "eslint-plugin-react": "^6.10.0",
     "gh-pages": "^1.0.0",
-    "jest": "^19.0.2",
+    "jest": "^22.4.3",
     "jest-bamboo-formatter": "0.0.3",
     "jest-serializer-path": "^0.1.14",
     "jscs": "^2.3.5",
