--- test/openCloseTest.js
@@ -1,146 +1,0 @@
-/* Copyright (c) 2010 - 2017, Nordic Semiconductor ASA
- *
- * All rights reserved.
- *
- * Use in source and binary forms, redistribution in binary form only, with
- * or without modification, are permitted provided that the following conditions
- * are met:
- *
- * 1. Redistributions in binary form, except as embedded into a Nordic
- *    Semiconductor ASA integrated circuit in a product or a software update for
- *    such product, must reproduce the above copyright notice, this list of
- *    conditions and the following disclaimer in the documentation and/or other
- *    materials provided with the distribution.
- *
- * 2. Neither the name of Nordic Semiconductor ASA nor the names of its
- *    contributors may be used to endorse or promote products derived from this
- *    software without specific prior written permission.
- *
- * 3. This software, with or without modification, must only be used with a Nordic
- *    Semiconductor ASA integrated circuit.
- *
- * 4. Any software provided in binary form under this license must not be reverse
- *    engineered, decompiled, modified and/or disassembled.
- *
- * THIS SOFTWARE IS PROVIDED BY NORDIC SEMICONDUCTOR ASA "AS IS" AND ANY EXPRESS OR
- * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
- * MERCHANTABILITY, NONINFRINGEMENT, AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL NORDIC SEMICONDUCTOR ASA OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR
- * TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
- * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-'use strict';
-
-const { grabAdapter, releaseAdapter, setupAdapter, outcome } = require('./setup');
-
-const debug = require('debug')('debug');
-const error = require('debug')('test:error');
-const testOutcome = require('debug')('test:outcome');
-
-const centralDeviceAddress = 'FF:11:22:33:AA:BE';
-const centralDeviceAddressType = 'BLE_GAP_ADDR_TYPE_RANDOM_STATIC';
-
-async function startScan(adapter, timeout) {
-    const scanParameters = {
-        active: true,
-        interval: 100,
-        window: 20,
-        timeout,
-    };
-
-    return new Promise((resolve, reject) => {
-        adapter.startScan(scanParameters, err => {
-            if (err) {
-                reject(err);
-                return;
-            }
-
-            resolve();
-        });
-    });
-}
-
-function runTests(numberOfIterations) {
-    const expectedNumberOfScanReports = 2;
-
-    let adapterSn;
-    let openCloseIterations = 0;
-    let programDevice = true;
-    let sdApiVersion;
-    const requiredNumberOfIterations = numberOfIterations || Number.MAX_SAFE_INTEGER;
-
-    const oneIteration = async () => {
-        debug(`Open/close iteration #${openCloseIterations} starting.`);
-
-        const adapterToUse = await grabAdapter(adapterSn, { programDevice });
-
-        // Program/check for correct firmware only once
-        programDevice = false;
-        adapterSn = adapterToUse.state.serialNumber;
-        sdApiVersion = adapterToUse.driver.NRF_SD_BLE_API_VERSION;
-
-        await setupAdapter(adapterToUse, 'central', 'central', centralDeviceAddress, centralDeviceAddressType);
-
-        let scanReportsReceived = 0;
-
-        const scanReportsReceivedPromise = new Promise(scanReportReceivedResolve => {
-            const deviceDiscoveredEvent = () => {
-                scanReportsReceived += 1;
-
-                if (scanReportsReceived >= expectedNumberOfScanReports) {
-                    debug(`Received ${scanReportsReceived} scan reports.`);
-                    adapterToUse.removeListener('deviceDiscovered', deviceDiscoveredEvent);
-                    scanReportReceivedResolve();
-                }
-            };
-
-            adapterToUse.on('deviceDiscovered', deviceDiscoveredEvent);
-        });
-
-        await startScan(adapterToUse, 2000);
-        await outcome([scanReportsReceivedPromise], 1500, `Failed waiting for ${expectedNumberOfScanReports} scan reports.`);
-
-        await new Promise((resolve, reject) => {
-            adapterToUse.stopScan(stopScanErr => {
-                if (stopScanErr) {
-                    reject(stopScanErr);
-                    return;
-                }
-
-                resolve();
-            });
-        });
-
-        await releaseAdapter(adapterSn);
-        testOutcome(`Open/close iteration #${openCloseIterations} of ${requiredNumberOfIterations} complete.`);
-        openCloseIterations += 1;
-    };
-
-    return new Promise(async (_, reject) => {
-        try {
-            while (openCloseIterations < requiredNumberOfIterations) {
-                // eslint-disable-next-line no-await-in-loop
-                await oneIteration();
-
-                if (sdApiVersion === 2) {
-                    // eslint-disable-next-line no-await-in-loop
-                    await new Promise(resolve => setTimeout(resolve, 250));
-                }
-            }
-        } catch (iterationErr) {
-            reject(iterationErr);
-        }
-    });
-}
-
-// Should run until error occurs
-// This test expects an external device to advertise at least 2 packets in the test scans for advertisement packets
-runTests(1000).catch(failure => {
-    error(failure);
-    process.exit(-1);
-});
